requirements.txt
fastapi
uvicorn
aiosmtplib
dnspython
email-validator
